Extract formatTime util
