+++ src/lib/media/video/formatTime.ts
@@ -1,0 +1,16 @@
+/**
+ * Formats a duration in seconds as minutes and seconds for video controls and
+ * related UI.
+ */
+export function formatTime(time: number) {
+  if (isNaN(time)) {
+    return '--'
+  }
+
+  time = Math.round(time)
+
+  const minutes = Math.floor(time / 60)
+  const seconds = String(time % 60).padStart(2, '0')
+
+  return `${minutes}:${seconds}`
+}
